requirements.txt
python-telegram-bot==13.7  # Hoặc phiên bản mới nhất bạn đang dùng
pandas
openpyxl
psycopg2-binary